+++ src/components/Sliders/Hue/HueSlider.tsx
