+++ src/index.ts
@@ -13,9 +13,9 @@ export { Panel3 } from '@panels/Panel3/Panel3';
 export { Panel4 } from '@panels/Panel4';
 export { Panel5 } from '@panels/Panel5';
 
-export { ExtraThumb } from '@panels/Panel3/ExtraThumb';
+export { HueSlider } from '@sliders/Hue/HueSlider';
+export { HueCircular } from '@sliders/Hue/HueCircular';
 
-export { HueSlider } from '@sliders/HSB/HueSlider';
 export { SaturationSlider } from '@sliders/HSB/SaturationSlider';
 export { BrightnessSlider } from '@sliders/HSB/BrightnessSlider';
 
@@ -23,7 +23,6 @@ export { LuminanceSlider } from '@sliders/HSL/LuminanceSlider';
 export { HSLSaturationSlider } from '@sliders/HSL/HSLSaturationSlider';
 
 export { OpacitySlider } from '@sliders/OpacitySlider';
-export { HueCircular } from '@sliders/HueCircular';
 
 export { RedSlider } from '@sliders/RGB/RedSlider';
 export { GreenSlider } from '@sliders/RGB/GreenSlider';
@@ -31,9 +30,8 @@ export { BlueSlider } from '@sliders/RGB/BlueSlider';
 
 export { Preview } from './components/Preview';
 export { PreviewText } from './components/PreviewText';
-
 export { InputWidget } from './components/InputWidget/InputWidget';
-
 export { Swatches } from './components/Swatches';
+export { ExtraThumb } from '@panels/Panel3/ExtraThumb';
 
 export * from '@types';
